voice-chatbot-app/requirements.txt
streamlit
azure-cognitiveservices-speech
openai
pydub
numpy
requests
python-dotenv
